Complete Projects page
I added the modify, create and delete functionality for projects with popovers and helper functions. Remember the whole logic runs on client side! So I am aware that thats not production code but I wanted to use tRPC and I am not so experienced with it so, I hope its fine